+++ src/main/java/com/invadermonky/bakedenchants/BakedEnchants.java
@@ -7,6 +7,7 @@ import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.fml.common.Loader;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import org.apache.logging.log4j.LogManager;
@@ -17,10 +18,12 @@ import org.apache.logging.log4j.Logger;
 public class BakedEnchants {
     public static final String MOD_ID = Tags.MOD_ID;
     public static final String MOD_NAME = Tags.MOD_NAME;
-    public static final String MOD_VERSION = "2.0.2";
-    public static final String DEPENDENCIES = "required-after:mixinbooter";
+    public static final String MOD_VERSION = "2.1.0";
+    public static final String DEPENDENCIES = "required-after:mixinbooter;after:crafttweaker";
 
     public static final Logger LOGGER = LogManager.getLogger(MOD_NAME);
+
+    public static boolean crafttweakerLoaded = Loader.isModLoaded("crafttweaker");
 
     @SubscribeEvent
     public static void registerRecipes(RegistryEvent.Register<IRecipe> event) {
@@ -29,8 +32,15 @@ public class BakedEnchants {
 
     @Mod.EventHandler
     public void preInit(FMLPreInitializationEvent event) {
-        if (Loader.isModLoaded("crafttweaker")) {
+        if (crafttweakerLoaded) {
             MinecraftForge.EVENT_BUS.register(new CTIntegration());
+        }
+    }
+
+    @Mod.EventHandler
+    public void postInit(FMLPostInitializationEvent event) {
+        if (crafttweakerLoaded) {
+            CTIntegration.postInit();
         }
     }
 }