+++ src/main/java/com/invadermonky/bakedenchants/compat/crafttweaker/CTBakedEnchants.java
@@ -6,21 +6,69 @@ import com.invadermonky.bakedenchants.handler.BakedEnchantmentRecipe;
 import crafttweaker.CraftTweakerAPI;
 import crafttweaker.annotations.ZenRegister;
 import crafttweaker.api.enchantments.IEnchantment;
+import crafttweaker.api.item.IIngredient;
 import crafttweaker.api.item.IItemStack;
 import crafttweaker.api.minecraft.CraftTweakerMC;
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.item.ItemStack;
+import net.minecraft.item.crafting.IRecipe;
+import net.minecraft.item.crafting.Ingredient;
+import net.minecraft.item.crafting.ShapedRecipes;
+import net.minecraft.item.crafting.ShapelessRecipes;
+import net.minecraft.util.NonNullList;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.fml.common.registry.ForgeRegistries;
 import stanhebben.zenscript.annotations.ZenClass;
 import stanhebben.zenscript.annotations.ZenMethod;
+
+import java.util.Arrays;
 
 @ZenRegister
 @ZenClass("mods." + BakedEnchants.MOD_ID + ".BakedEnchants")
 public class CTBakedEnchants {
 
     @ZenMethod
-    public static void add(IItemStack iStack, IEnchantment... iEnchantments) {
+    public static void addShapedRecipe(IItemStack output, IIngredient[][] ingredients) {
+        addShapedRecipe(CraftTweakerMC.getItemStack(output).getItem().getRegistryName().getPath(), output, ingredients);
+    }
+
+    @ZenMethod
+    public static void addShapedRecipe(String recipeName, IItemStack output, IIngredient[][] ingredients) {
+        int height = ingredients.length;
+        int width = Arrays.stream(ingredients).mapToInt(ing -> ing.length).max().orElse(0);
+        NonNullList<Ingredient> inputs = NonNullList.create();
+        Arrays.stream(ingredients).forEach(inputRow -> Arrays.stream(inputRow).map(CraftTweakerMC::getIngredient).forEach(inputs::add));
+        IRecipe recipe = new ShapedRecipes("", width, height, inputs, CraftTweakerMC.getItemStack(output));
+        ResourceLocation loc = new ResourceLocation(BakedEnchants.MOD_ID, recipeName);
+        int i = 0;
+        while (CTIntegration.craftingRecipeCache.containsKey(loc)) {
+            loc = new ResourceLocation(BakedEnchants.MOD_ID, recipeName + "_alt" + i++);
+        }
+        recipe.setRegistryName(loc);
+        CTIntegration.cacheCraftingTableRecipe(recipe);
+    }
+
+    @ZenMethod
+    public static void addShapelessRecipe(IItemStack output, IIngredient[] ingredients) {
+        addShapelessRecipe(CraftTweakerMC.getItemStack(output).getItem().getRegistryName().getPath(), output, ingredients);
+    }
+
+    @ZenMethod
+    public static void addShapelessRecipe(String recipeName, IItemStack output, IIngredient[] ingredients) {
+        NonNullList<Ingredient> inputs = NonNullList.create();
+        Arrays.stream(ingredients).map(CraftTweakerMC::getIngredient).forEach(inputs::add);
+        IRecipe recipe = new ShapelessRecipes("", CraftTweakerMC.getItemStack(output), inputs);
+        ResourceLocation loc = new ResourceLocation(BakedEnchants.MOD_ID, recipeName);
+        int i = 0;
+        while (CTIntegration.craftingRecipeCache.containsKey(loc)) {
+            loc = new ResourceLocation(BakedEnchants.MOD_ID, recipeName + "_alt" + i++);
+        }
+        recipe.setRegistryName(loc);
+        CTIntegration.cacheCraftingTableRecipe(recipe);
+    }
+
+    @ZenMethod
+    public static void addBakedEnchantment(IItemStack iStack, IEnchantment... iEnchantments) {
         ItemStack stack = CraftTweakerMC.getItemStack(iStack);
         if (stack.isEmpty()) {
             CraftTweakerAPI.logError("IItemStack cannot be empty");